Melhorias diversas; Fechamento da primeira versão.

+++ src/main/java/br/com/emerion/config/JwtTokenUtil.java
@@ -12,7 +12,7 @@ import io.jsonwebtoken.Jwts;
 public class JwtTokenUtil implements Serializable {
 
 	private static final long serialVersionUID = -2550185165626007488L;
-	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+	public static final long JWT_TOKEN_VALIDITY = (long)5 * 60 * 60;
 
 	public String getUsernameFromToken(String token) {
 		return getClaimFromToken(token, Claims::getSubject);
